Move LoginFlowResult to login/flowResult
