+++ src/frontend/src/flows/confirmRegister.ts
@@ -2,7 +2,10 @@ import { html, render } from "lit-html";
 import { displayUserNumber } from "./displayUserNumber";
 import { displayError } from "../components/displayError";
 import { setUserNumber } from "../utils/userNumber";
-import { apiResultToLoginFlowResult, LoginFlowResult } from "./loginUnknown";
+import {
+  apiResultToLoginFlowResult,
+  LoginFlowResult,
+} from "./login/flowResult";
 import { Challenge } from "../../generated/internet_identity_types";
 import { WebAuthnIdentity } from "@dfinity/identity";
 import getProofOfWork from "../crypto/pow";
