+++ src/frontend/src/flows/loginUnknown.ts
@@ -1,5 +1,5 @@
 import { render, html } from "lit-html";
-import { IIConnection, ApiResult } from "../utils/iiConnection";
+import { IIConnection } from "../utils/iiConnection";
 import { parseUserNumber, setUserNumber } from "../utils/userNumber";
 import { withLoader } from "../components/loader";
 import { register } from "./register";
@@ -9,6 +9,10 @@ import { navbar } from "../components/navbar";
 import { UserIntent, authenticateUnknownIntent } from "../utils/userIntent";
 import { useRecovery } from "./recovery/useRecovery";
 import { registerDisabled } from "./registerDisabled";
+import {
+  apiResultToLoginFlowResult,
+  LoginFlowResult,
+} from "./login/flowResult";
 
 const pageContent = (userIntent: UserIntent) => html` <style>
     #registerUserNumber:focus {
@@ -71,19 +75,6 @@ const pageContent = (userIntent: UserIntent) => html` <style>
           </div>`}
   </div>
   ${navbar}`;
-
-export type LoginFlowResult =
-  | {
-      tag: "ok";
-      userNumber: bigint;
-      connection: IIConnection;
-    }
-  | {
-      tag: "err";
-      title: string;
-      message: string;
-      detail?: string;
-    };
 
 export const loginUnknown = async (
   userIntent: UserIntent
@@ -184,66 +175,4 @@ const initLinkDevice = () => {
     const userNumber = parseUserNumber(userNumberInput.value);
     addDeviceUserNumber(userNumber);
   };
-};
-
-export const apiResultToLoginFlowResult = (result: ApiResult): LoginFlowResult => {
-  switch (result.kind) {
-    case "loginSuccess": {
-      return {
-        tag: "ok",
-        userNumber: result.userNumber,
-        connection: result.connection,
-      };
-    }
-    case "authFail": {
-      return {
-        tag: "err",
-        title: "Failed to authenticate",
-        message:
-          "We failed to authenticate you using your security device. If this is the first time you're trying to log in with this device, you have to add it as a new device first.",
-        detail: result.error.message,
-      };
-    }
-    case "unknownUser": {
-      return {
-        tag: "err",
-        title: "Unknown Identity Anchor",
-        message: `Failed to find an identity for the Identity Anchor ${result.userNumber}. Please check your Identity Anchor and try again.`,
-        detail: "",
-      };
-    }
-    case "apiError": {
-      return {
-        tag: "err",
-        title: "We couldn't reach Internet Identity",
-        message:
-          "We failed to call the Internet Identity service, please try again.",
-        detail: result.error.message,
-      };
-    }
-    case "registerNoSpace": {
-      return {
-        tag: "err",
-        title: "Failed to register",
-        message:
-          "Failed to register with Internet Identity, because there is no space left at the moment. We're working on increasing the capacity.",
-      };
-    }
-    case "badChallenge": {
-      return {
-        tag: "err",
-        title: "Failed to register",
-        message:
-          "Failed to register with Internet Identity, because the CAPTCHA challenge wasn't successful",
-      };
-    }
-    case "seedPhraseFail": {
-      return {
-        tag: "err",
-        title: "Invalid Seed Phrase",
-        message:
-          "Failed to recover using this seedphrase. Did you enter it correctly?",
-      };
-    }
-  }
 };
